+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.2.8
 * Magic Dagger : Added small vfx on use.
+* Thermite Flare : Fixed error with the increased damage effect.
 * Slightly faster actor collision checking for Explosive Spear and Whimsical Star.
 
 ### v1.2.7